refactor: Move rental detail related files to appropriate packages (#145)
- Move screens → `presentation.rentaldetail`
  - `OwnerRentalDetailScreen.kt`, `RenterRentalDetailScreen.kt`
- Move UI content → `content` sub-package
  - e.g., `OwnerPaidContent.kt`, `RenterRequestContent.kt`
- Move photo-related screens → `photobeforerent` / `photobeforereturn`
  - e.g., `PhotoBeforeRentScreen.kt`, `PhotoBeforeReturnScreen.kt`
- Move & rename model → `domain.rental.model`
  - `RentingStatus.kt` → `RentingStatusModel.kt`
- Other updates:
  - Update imports
  - Remove direct screen imports in `RentalDetailRoute.kt`
  - Rename preview → `OwnerRentalDetailScreenPreview`
