+++ app/src/main/java/com/example/rentit/data/rental/mapper/RentalDetailStatusMapper.kt
@@ -7,7 +7,7 @@ import com.example.rentit.common.util.daysFromToday
 import com.example.rentit.data.rental.dto.RentalDetailResponseDto
 import com.example.rentit.common.uimodel.RentalSummaryUiModel
 import com.example.rentit.domain.rental.model.RentalDetailStatusModel
-import com.example.rentit.presentation.rentaldetail.model.RentingStatus
+import com.example.rentit.domain.rental.model.RentingStatus
 
 @RequiresApi(Build.VERSION_CODES.O)
 fun RentalDetailResponseDto.toModel(): RentalDetailStatusModel {
