+++ app/src/main/java/com/example/rentit/domain/rental/model/RentingStatusModel.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.presentation.rentaldetail.model
+package com.example.rentit.domain.rental.model
 
 import androidx.annotation.StringRes
 import androidx.compose.ui.graphics.Color
