+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/photobeforereturn/PhotoBeforeReturnSideEffect.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.presentation.rentaldetail.renter.photobeforereturn
+package com.example.rentit.presentation.rentaldetail.photobeforereturn
 
 sealed class PhotoBeforeReturnSideEffect {
     data object PopBackToRentalDetail : PhotoBeforeReturnSideEffect()
